Mixed-precision wire format: float64 positions + float32 velocities
Fixes the trail wobble regression caused by Phase 1's float32 positions.

Symptom: at high fidelity (high DP853 N, or low K for fixed-step), outer
planets' orbit planes appeared to jitter frame to frame. Affected all
three integrators (Euler, RK4, DP853) → the cause was in shared
infrastructure, not the integrator logic.

Root cause (confirmed by reverting wire to full float64 as a diagnostic
and observing the wobble disappear): float32's ~540 km position
quantization at Neptune's 4.5×10¹² m radius dominated per-sample Z
motion (~610 km at hours-unit K=1). Per-sample motion shrinks with
higher fidelity, so the quantization/motion ratio worsens monotonically
— matching the observed pattern exactly.

Fix: mixed precision on the wire. Positions stay float64 (rendered
directly, per-pixel sensitive). Velocities downgraded to float32
(used only as Hermite tangents and Keplerian inputs that damp
precision loss by ~5 orders of magnitude — invisible). 75% of full
float64 bytes; 1.5× Phase 1's bytes.

Updated wire-size ceilings (vs design doc's original 2 MB / 6 MB):
- Default tier (Euler K=1 highest): ~3 MB compressed
- DP853 tier (N=15000 highest): ~4.5 MB compressed
All within rate limiter sizing.

Doc updates: ARCHITECTURE.md wire-format paragraph + DP853 emission
section + Phase 3 numbers. Design doc gets a "Postscript: float32 →
mixed precision" section with the math and the lesson on validating
precision changes against the full body × scale × fidelity matrix.

Tests: SimulationTest 9/9, BinaryResponseSerializerTest 4/4
(mixed-precision assertions), parseBinaryChunk.test.ts 5/5. Frontend
build + lint clean. ChunkSizeBenchmark confirms new wire sizes.

+++ backend/src/test/java/personal/spacesim/utils/serializers/BinaryResponseSerializerTest.java
@@ -113,13 +113,16 @@ class BinaryResponseSerializerTest {
 
         assertEquals(1, buf.getInt(), "timestepCount");
 
-        // Per-timestep — positions+velocities are float32 (small integers roundtrip exactly).
+        // Per-timestep — positions are float64 (precision matters for outer
+        // planet orbit-plane rendering), velocities are float32 (their
+        // precision-loss path damps by ~5 orders of magnitude before
+        // anything visible).
         assertEquals(expectedMillis, buf.getLong(), "timestamp millis");
-        // Earth
-        assertEquals(1.0f, buf.getFloat()); assertEquals(2.0f, buf.getFloat()); assertEquals(3.0f, buf.getFloat());
+        // Earth — position float64, velocity float32.
+        assertEquals(1.0, buf.getDouble()); assertEquals(2.0, buf.getDouble()); assertEquals(3.0, buf.getDouble());
         assertEquals(4.0f, buf.getFloat()); assertEquals(5.0f, buf.getFloat()); assertEquals(6.0f, buf.getFloat());
         // Moon
-        assertEquals(7.0f, buf.getFloat()); assertEquals(8.0f, buf.getFloat()); assertEquals(9.0f, buf.getFloat());
+        assertEquals(7.0, buf.getDouble()); assertEquals(8.0, buf.getDouble()); assertEquals(9.0, buf.getDouble());
         assertEquals(10.0f, buf.getFloat()); assertEquals(11.0f, buf.getFloat()); assertEquals(12.0f, buf.getFloat());
 
         // Buffer should be exactly consumed
